# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ResQAI incident response API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type IncidentStatus = typeof IncidentStatus[keyof typeof IncidentStatus];


export const IncidentStatus = {
  awaiting_response: 'awaiting_response',
  in_progress: 'in_progress',
  resolved: 'resolved',
} as const;

export type PriorityLevel = typeof PriorityLevel[keyof typeof PriorityLevel];


export const PriorityLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export type AnalysisStatus = typeof AnalysisStatus[keyof typeof AnalysisStatus];


export const AnalysisStatus = {
  pending: 'pending',
  complete: 'complete',
  failed: 'failed',
} as const;

export type DisasterType = typeof DisasterType[keyof typeof DisasterType];


export const DisasterType = {
  flood: 'flood',
  earthquake: 'earthquake',
  cyclone: 'cyclone',
  landslide: 'landslide',
  fire: 'fire',
  other: 'other',
} as const;

export interface Incident {
  id: string;
  /** @nullable */
  reporterName: string | null;
  description: string;
  disasterType: DisasterType;
  locationDescription: string;
  /**
     * @minimum -90
     * @maximum 90
     * @nullable
     */
  latitude: number | null;
  /**
     * @minimum -180
     * @maximum 180
     * @nullable
     */
  longitude: number | null;
  /** @minimum 0 */
  peopleAffected: number;
  /** @nullable */
  imageUrl: string | null;
  analysisStatus: AnalysisStatus;
  /** @nullable */
  incidentType: string | null;
  /**
     * @minimum 0
     * @maximum 10
     * @nullable
     */
  severity: number | null;
  /** @nullable */
  vulnerablePerson: boolean | null;
  /**
     * @minimum 0
     * @maximum 10
     * @nullable
     */
  immediateThreat: number | null;
  /**
     * @minimum 0
     * @maximum 10
     * @nullable
     */
  accessibility: number | null;
  /**
     * @minimum 0
     * @maximum 10
     * @nullable
     */
  timeCriticality: number | null;
  /**
     * @minimum 0
     * @maximum 10
     * @nullable
     */
  priorityScore: number | null;
  priorityLevel: PriorityLevel | null;
  /** @nullable */
  aiSummary: string | null;
  status: IncidentStatus;
  isDemo: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface IncidentInput {
  /**
     * @minLength 1
     * @nullable
     */
  reporterName?: string | null;
  /** @minLength 10 */
  description: string;
  disasterType: DisasterType;
  /** @minLength 1 */
  locationDescription: string;
  /**
     * @minimum -90
     * @maximum 90
     * @nullable
     */
  latitude?: number | null;
  /**
     * @minimum -180
     * @maximum 180
     * @nullable
     */
  longitude?: number | null;
  /** @minimum 0 */
  peopleAffected: number;
  /** @nullable */
  imageUrl?: string | null;
}

export interface IncidentStatusUpdate {
  status: IncidentStatus;
}

export interface DashboardIncident {
  id: string;
  title: string;
  description: string;
  disasterType: DisasterType;
  locationDescription: string;
  /** @nullable */
  latitude: number | null;
  /** @nullable */
  longitude: number | null;
  severity: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  /** @minimum 0 */
  peopleAffected: number;
  source: string;
  timestamp: string;
  accessibility: number;
  status: IncidentStatus;
  priorityScore: number;
  priorityLevel: PriorityLevel;
  incidentType: string;
  aiSummary: string;
  isDemo: boolean;
}

export interface DashboardStats {
  activeIncidents: number;
  criticalIncidents: number;
  affectedPopulation: number;
  activeOperations: number;
  incomingReports: number;
}

export interface DashboardRecommendation {
  id: string;
  priority: PriorityLevel;
  action: string;
  reason: string;
  targetLocation: string;
  requiredResource: string;
  status: string;
}

export interface DashboardResponse {
  incidents: DashboardIncident[];
  stats: DashboardStats;
  summary: string;
  recommendations: DashboardRecommendation[];
  generatedAt: string;
  isDemo: boolean;
}

export interface ErrorResponse {
  error: string;
}