# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ResQAI incident response API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: incidents
    description: Disaster incident reporting and response operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [incidents]
      summary: Get the emergency operations dashboard
      description: Returns incident statistics, prioritized rescue zones, an operational summary, and response recommendations.
      responses:
        "200":
          description: Emergency operations dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardResponse"
  /incidents:
    get:
      operationId: listIncidents
      tags: [incidents]
      summary: List incidents
      description: Returns incidents ordered from newest to oldest.
      responses:
        "200":
          description: Incident list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Incident"
    post:
      operationId: createIncident
      tags: [incidents]
      summary: Create an incident report
      description: Stores a citizen or field incident report for later analysis.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentInput"
      responses:
        "201":
          description: Created incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "400":
          description: Invalid incident input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /incidents/{id}:
    get:
      operationId: getIncident
      tags: [incidents]
      summary: Get an incident
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      responses:
        "200":
          description: Incident details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "400":
          description: Invalid incident ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Incident not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /incidents/{id}/status:
    patch:
      operationId: updateIncidentStatus
      tags: [incidents]
      summary: Update incident status
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentStatusUpdate"
      responses:
        "200":
          description: Updated incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "400":
          description: Invalid incident ID or status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Incident not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    IncidentId:
      name: id
      in: path
      required: true
      description: Incident UUID
      schema:
        type: string
        format: uuid
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    IncidentStatus:
      type: string
      enum:
        - awaiting_response
        - in_progress
        - resolved
    PriorityLevel:
      type: string
      enum:
        - low
        - medium
        - high
        - critical
    AnalysisStatus:
      type: string
      enum:
        - pending
        - complete
        - failed
    DisasterType:
      type: string
      enum:
        - flood
        - earthquake
        - cyclone
        - landslide
        - fire
        - other
    Incident:
      type: object
      required:
        - id
        - reporterName
        - description
        - disasterType
        - locationDescription
        - latitude
        - longitude
        - peopleAffected
        - imageUrl
        - analysisStatus
        - incidentType
        - severity
        - vulnerablePerson
        - immediateThreat
        - accessibility
        - timeCriticality
        - priorityScore
        - priorityLevel
        - aiSummary
        - status
        - isDemo
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        reporterName:
          type: ["string", "null"]
        description:
          type: string
        disasterType:
          $ref: "#/components/schemas/DisasterType"
        locationDescription:
          type: string
        latitude:
          type: ["number", "null"]
          minimum: -90
          maximum: 90
        longitude:
          type: ["number", "null"]
          minimum: -180
          maximum: 180
        peopleAffected:
          type: integer
          minimum: 0
        imageUrl:
          type: ["string", "null"]
          format: uri
        analysisStatus:
          $ref: "#/components/schemas/AnalysisStatus"
        incidentType:
          type: ["string", "null"]
        severity:
          type: ["number", "null"]
          minimum: 0
          maximum: 10
        vulnerablePerson:
          type: ["boolean", "null"]
        immediateThreat:
          type: ["number", "null"]
          minimum: 0
          maximum: 10
        accessibility:
          type: ["number", "null"]
          minimum: 0
          maximum: 10
        timeCriticality:
          type: ["number", "null"]
          minimum: 0
          maximum: 10
        priorityScore:
          type: ["number", "null"]
          minimum: 0
          maximum: 10
        priorityLevel:
          oneOf:
            - $ref: "#/components/schemas/PriorityLevel"
            - type: "null"
        aiSummary:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/IncidentStatus"
        isDemo:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    IncidentInput:
      type: object
      required:
        - description
        - disasterType
        - locationDescription
        - peopleAffected
      properties:
        reporterName:
          type: ["string", "null"]
          minLength: 1
        description:
          type: string
          minLength: 10
        disasterType:
          $ref: "#/components/schemas/DisasterType"
        locationDescription:
          type: string
          minLength: 1
        latitude:
          type: ["number", "null"]
          minimum: -90
          maximum: 90
        longitude:
          type: ["number", "null"]
          minimum: -180
          maximum: 180
        peopleAffected:
          type: integer
          minimum: 0
        imageUrl:
          type: ["string", "null"]
          format: uri
    IncidentStatusUpdate:
      type: object
      required:
        - status
      properties:
        status:
          $ref: "#/components/schemas/IncidentStatus"
    DashboardIncident:
      type: object
      required:
        - id
        - title
        - description
        - disasterType
        - locationDescription
        - latitude
        - longitude
        - severity
        - confidence
        - peopleAffected
        - source
        - timestamp
        - accessibility
        - status
        - priorityScore
        - priorityLevel
        - incidentType
        - aiSummary
        - isDemo
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        disasterType:
          $ref: "#/components/schemas/DisasterType"
        locationDescription:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        severity:
          type: number
        confidence:
          type: number
          minimum: 0
          maximum: 1
        peopleAffected:
          type: integer
          minimum: 0
        source:
          type: string
        timestamp:
          type: string
          format: date-time
        accessibility:
          type: number
        status:
          $ref: "#/components/schemas/IncidentStatus"
        priorityScore:
          type: number
        priorityLevel:
          $ref: "#/components/schemas/PriorityLevel"
        incidentType:
          type: string
        aiSummary:
          type: string
        isDemo:
          type: boolean
    DashboardStats:
      type: object
      required:
        - activeIncidents
        - criticalIncidents
        - affectedPopulation
        - activeOperations
        - incomingReports
      properties:
        activeIncidents:
          type: integer
        criticalIncidents:
          type: integer
        affectedPopulation:
          type: integer
        activeOperations:
          type: integer
        incomingReports:
          type: integer
    DashboardRecommendation:
      type: object
      required:
        - id
        - priority
        - action
        - reason
        - targetLocation
        - requiredResource
        - status
      properties:
        id:
          type: string
        priority:
          $ref: "#/components/schemas/PriorityLevel"
        action:
          type: string
        reason:
          type: string
        targetLocation:
          type: string
        requiredResource:
          type: string
        status:
          type: string
    DashboardResponse:
      type: object
      required:
        - incidents
        - stats
        - summary
        - recommendations
        - generatedAt
        - isDemo
      properties:
        incidents:
          type: array
          items:
            $ref: "#/components/schemas/DashboardIncident"
        stats:
          $ref: "#/components/schemas/DashboardStats"
        summary:
          type: string
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/DashboardRecommendation"
        generatedAt:
          type: string
          format: date-time
        isDemo:
          type: boolean
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string