# lib/api-zod/src/generated/types/dashboardIncident.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ResQAI incident response API
 * OpenAPI spec version: 0.1.0
 */
import type { DisasterType } from './disasterType';
import type { IncidentStatus } from './incidentStatus';
import type { PriorityLevel } from './priorityLevel';

export interface DashboardIncident {
  id: string;
  title: string;
  description: string;
  disasterType: DisasterType;
  locationDescription: string;
  /** @nullable */
  latitude: number | null;
  /** @nullable */
  longitude: number | null;
  severity: number;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  /** @minimum 0 */
  peopleAffected: number;
  source: string;
  timestamp: Date;
  accessibility: number;
  status: IncidentStatus;
  priorityScore: number;
  priorityLevel: PriorityLevel;
  incidentType: string;
  aiSummary: string;
  isDemo: boolean;
}