# lib/api-zod/src/generated/types/dashboardRecommendation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ResQAI incident response API
 * OpenAPI spec version: 0.1.0
 */
import type { PriorityLevel } from './priorityLevel';

export interface DashboardRecommendation {
  id: string;
  priority: PriorityLevel;
  action: string;
  reason: string;
  targetLocation: string;
  requiredResource: string;
  status: string;
}