# lib/api-zod/src/generated/types/dashboardResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ResQAI incident response API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardIncident } from './dashboardIncident';
import type { DashboardRecommendation } from './dashboardRecommendation';
import type { DashboardStats } from './dashboardStats';

export interface DashboardResponse {
  incidents: DashboardIncident[];
  stats: DashboardStats;
  summary: string;
  recommendations: DashboardRecommendation[];
  generatedAt: Date;
  isDemo: boolean;
}