# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ResQAI incident response API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  activeIncidents: number;
  criticalIncidents: number;
  affectedPopulation: number;
  activeOperations: number;
  incomingReports: number;
}